Migraciones y Consultas

+++ app/Http/Controllers/PacientesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Pacientes;
+use App\Pacientes_nt;
 use App\Http\Requests\PacientesRequest;
 use Laracasts\Flash\flash;
 
@@ -60,7 +61,21 @@ class PacientesController extends Controller
                                          'telefono' => $request->telefono,
                                          'direccion' => $request->direccion,
                                          'titular' => $titular,
-                                         'institucion' => $request->institucion]);
+                                         'institucion' => $request->institucion,
+                                         'edad' => $request->edad,
+                                         'genero' => $request->genero]);
+
+            $pacientent=Pacientes_nt::create(['nombres' => $request->nombres,
+                                         'apellidos' => $request->apellidos,
+                                         'nacionalidad' => $request->nacionalidad,
+                                         'cedula' => $request->cedula,
+                                         'codigo_telf' => $request->codigo_telf,
+                                         'telefono' => $request->telefono,
+                                         'direccion' => $request->direccion,
+                                         'titular' => 'Si',
+                                         'edad' => $request->edad,
+                                         'genero' => $request->genero,
+                                         'id_paciente' => $paciente->id]);
             flash("EL REGISTRO HA SIDO EXITOSO!", 'success');            
         }
 
@@ -107,13 +122,9 @@ class PacientesController extends Controller
             flash("LA CÉDULA YA HA SIDO REGISTRADA!", 'error'); 
             return redirect()->route('pacientes.edit')->withInput();
         } else {
-            if ($request->titular!="Si") {
-                $titular="No";
-            } else {
-                $titular="Si";
-            }
-
+            
             $paciente=Pacientes::find($id);
+            $cedula=$paciente->cedula;
             $paciente->nombres=$request->nombres;
             $paciente->apellidos=$request->apellidos;
             $paciente->nacionalidad=$request->nacionalidad;
@@ -121,9 +132,23 @@ class PacientesController extends Controller
             $paciente->codigo_telf=$request->codigo_telf;
             $paciente->telefono=$request->telefono;
             $paciente->direccion=$request->direccion;
-            $paciente->titular=$titular;
             $paciente->institucion=$request->institucion;
+            $paciente->edad=$request->edad;
+            $paciente->genero=$request->genero;
             $paciente->save();
+
+            $pacientent=Pacientes_nt::where('cedula',$cedula)->first();
+            $pacientent->nombres=$request->nombres;
+            $pacientent->apellidos=$request->apellidos;
+            $pacientent->nacionalidad=$request->nacionalidad;
+            $pacientent->cedula=$request->cedula;
+            $pacientent->codigo_telf=$request->codigo_telf;
+            $pacientent->telefono=$request->telefono;
+            $pacientent->direccion=$request->direccion;
+            $pacientent->edad=$request->edad;
+            $pacientent->genero=$request->genero;
+            $pacientent->id_paciente=$id;
+            $pacientent->save();
 
             flash("LA ACTUALIZACIÓN HA SIDO EXITOSA!", 'success'); 
             return redirect()->route('pacientes.index');
@@ -139,14 +164,23 @@ class PacientesController extends Controller
      */
     public function destroy(Request $request)
     {
-        //falta validar la existencia de paciente en otras tablas
-
         $paciente=Pacientes::find($request->id);
-        if ($paciente->delete()) {
-         flash("LA ELIMINACIÓN DEL REGISTRO HA SIDO EXITOSA!", 'success'); 
-            
-        } else {
-            flash("DISCULPE, LA ELIMINACIÓN DEL REGISTRO NO HA SIDO EXITOSA!", 'error'); 
+        $pacientent=Pacientes_nt::where('id_paciente',$request->id)->get();
+        if($pacientent->consultas->exists()){
+            flash("DISCULPE, NO SE PUEDE ELIMINAR YA QUE SE ENCUENTRA EN REGISTROS DE CONSULTAS!", 'error'); 
+        }else{
+            if (count($pacientent)>1) {
+              flash("DISCULPE, NO SE PUEDE ELIMINAR YA QUE EXISTEN PACIENTES QUE SE ENCUENTRAN COMO BENEFICIARIOS DE ESTE PACIENTE!", 'error');   
+            } else {
+                
+                $pacientent->delete();
+                if ($paciente->delete()) {
+                 flash("LA ELIMINACIÓN DEL REGISTRO HA SIDO EXITOSA!", 'success'); 
+                    
+                } else {
+                    flash("DISCULPE, LA ELIMINACIÓN DEL REGISTRO NO HA SIDO EXITOSA!", 'error'); 
+                }
+            }
         }
         return redirect()->route('pacientes.index');
     }
